+++ helpers/dateTime.js
@@ -66,7 +66,7 @@ export function replaceArrowDatesInString(inString: string, replaceWith: string 
     repl = getTodaysDateAsArrowDate()
   }
   while (isScheduled(str)) {
-    str = str.replace(RE_PLUS_DATE, '').replace('>today', '').replace(/ {2,}/g, ' ').trim()
+    str = str.replace(RE_PLUS_DATE, '').replace('>today', '').replace(new RegExp(WEEK_NOTE_LINK)).replace(/ {2,}/g, ' ').trim()
   }
   return repl && repl.length > 0 ? `${str} ${repl}` : str
 }